backend/requirements.txt
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# python-dotenv==1.0.0
# openai==1.3.8
# tiktoken==0.5.2
# pydantic==2.5.0
# aiofiles==23.2.1
# python-multipart==0.0.6

fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
openai>=1.12.0
tiktoken>=0.6.0
pydantic>=2.6.0
aiofiles>=23.2.1
python-multipart>=0.0.6